requirements.txt
streamlit
pypdf
google-generativeai
python-docx
Pillow
